fixed regions and moved file with database functions outside of react folder
