+++ dao.js
@@ -1,7 +1,7 @@
 var AWS = require("aws-sdk");
 
 AWS.config.update({
-  region: "us-west-2",
+  region: "us-west-1",
   endpoint: "http://localhost:8000"
 });
